% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTDNAdata.R
\name{fixTDNAcorruption}
\alias{fixTDNAcorruption}
\title{Fix package corruption issues}
\usage{
fixTDNAcorruption(repo = "ianandersonlol/tdna")
}
\arguments{
\item{repo}{Repository path for reinstallation. Default: "ianandersonlol/tdna".}
}
\value{
TRUE if successful
}
\description{
This function reinstalls the package to fix corruption issues.
}
