% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTDNAlines.R
\name{plotTDNAlines_old}
\alias{plotTDNAlines_old}
\title{Legacy version of the plotting function}
\usage{
plotTDNAlines_old(
  gene,
  show_introns = TRUE,
  show_all_features = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{gene}{A character string specifying the Arabidopsis gene ID}

\item{show_introns}{Whether to show introns}

\item{show_all_features}{Whether to show all features}

\item{interactive}{Whether to create an interactive plot}
}
\description{
Legacy version of the plotting function
}
\keyword{internal}
