% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTDNAdata.R
\name{verify_tdna_data}
\alias{verify_tdna_data}
\title{Check if T-DNA data is valid and reload if necessary}
\usage{
verify_tdna_data()
}
\description{
Check if T-DNA data is valid and reload if necessary
}
\keyword{internal}
