% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTDNAlines.R
\name{getTDNAlines}
\alias{getTDNAlines}
\title{return TDNA lines that are found in gene of interest}
\usage{
getTDNAlines(gene)
}
\arguments{
\item{gene}{A number.}
}
\description{
return TDNA lines that are found in gene of interest
}
