% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTDNAlines.R
\name{plotTDNAlines}
\alias{plotTDNAlines}
\title{Plot T-DNA insertions in a gene}
\usage{
plotTDNAlines(gene, data_env = .tdna_env)
}
\arguments{
\item{gene}{Gene identifier.}
\item{data_env}{Environment with the loaded data.}
}
\description{
Create a simple plot showing the gene's coding sequence segments and labels for confirmed homozygous T-DNA insertions.
}
